# puller: fix a memory leak bug in the frontier